requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
notebooklm-py>=0.1.0
gspread>=6.0.0
google-auth>=2.28.0
pydantic>=2.6.0
aiofiles>=23.2.0
slowapi>=0.1.9
tenacity>=8.2.0
